Make it work for CRI-O
