+++ pkg/util/trivy/overlayfs.go
@@ -10,233 +10,101 @@
 package trivy
 
 import (
+	"context"
 	"errors"
 	"fmt"
-	"io/fs"
-	"os"
-	"path"
-	"sort"
-	"syscall"
 
-	"golang.org/x/sys/unix"
+	workloadmeta "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
+	"github.com/DataDog/datadog-agent/pkg/sbom"
+	"github.com/DataDog/datadog-agent/pkg/util/log"
+	"github.com/aquasecurity/trivy/pkg/fanal/applier"
+	local "github.com/aquasecurity/trivy/pkg/fanal/artifact/container"
+	ftypes "github.com/aquasecurity/trivy/pkg/fanal/types"
+	v1 "github.com/google/go-containerregistry/pkg/v1"
 )
 
-// whiteoutCharDev is defined as zero and is not const only for testing as it
-// is not allowed to mknod a 0/0 char dev in userns.
-var whiteoutCharDev uint64 // = 0
-
-var whiteout *fs.DirEntry
-
-type filesystem struct {
-	layers []string
+type fakeContainer struct {
+	layerIDs   []string
+	imgMeta    *workloadmeta.ContainerImageMetadata
+	layerPaths []string
 }
 
-type file struct {
-	*os.File
-	ofs  filesystem
-	fi   fs.FileInfo
-	name string
-}
-
-// NewFS returns a fs.ReadDirFS consisting of merging the given layer paths.
-func NewFS(layers []string) interface {
-	fs.FS
-	fs.ReadDirFS
-	fs.StatFS
-} {
-	return &filesystem{layers[:]}
-}
-
-// Open implements fs.StatFS.
-func (ofs filesystem) Stat(name string) (fs.FileInfo, error) {
-	name = path.Join("/", name)[1:]
-	if name == "" {
-		name = "."
+func (c *fakeContainer) LayerByDiffID(hash string) (ftypes.LayerPath, error) {
+	for i, layer := range c.layerIDs {
+		diffID, _ := v1.NewHash(layer)
+		if diffID.String() == hash {
+			return ftypes.LayerPath{
+				DiffID: diffID.String(),
+				Path:   c.layerPaths[i],
+				Digest: c.imgMeta.Layers[i].Digest,
+			}, nil
+		}
 	}
-	_, fi, err := ofs.stat(name)
-	return fi, err
+	return ftypes.LayerPath{}, errors.New("not found")
 }
 
-// Open implements fs.FS.
-func (ofs filesystem) Open(name string) (fs.File, error) {
-	name = path.Join("/", name)[1:]
-	layerIndex, fi, err := ofs.stat(name)
+func (c *fakeContainer) LayerByDigest(hash string) (ftypes.LayerPath, error) {
+	for i, layer := range c.layerIDs {
+		diffID, _ := v1.NewHash(layer)
+		if hash == c.imgMeta.Layers[i].Digest {
+			return ftypes.LayerPath{
+				DiffID: diffID.String(),
+				Path:   c.layerPaths[i],
+				Digest: c.imgMeta.Layers[i].Digest,
+			}, nil
+		}
+	}
+	return ftypes.LayerPath{}, errors.New("not found")
+}
+
+func (c *fakeContainer) Layers() (layers []ftypes.LayerPath) {
+	for i, layer := range c.layerIDs {
+		diffID, _ := v1.NewHash(layer)
+		layers = append(layers, ftypes.LayerPath{
+			DiffID: diffID.String(),
+			Path:   c.layerPaths[i],
+			Digest: c.imgMeta.Layers[i].Digest,
+		})
+	}
+
+	return layers
+}
+
+func (c *Collector) scanOverlayFS(ctx context.Context, layers []string, ctr ftypes.Container, imgMeta *workloadmeta.ContainerImageMetadata, scanOptions sbom.ScanOptions) (sbom.Report, error) {
+	cache, err := c.getCache()
 	if err != nil {
-		err.(*os.PathError).Op = "open"
 		return nil, err
 	}
-	f, err := os.Open(ofs.path(layerIndex, name))
+
+	if cache == nil {
+		return nil, errors.New("failed to get cache for scan")
+	}
+
+	containerArtifact, err := local.NewArtifact(ctr, cache, NewFSWalker(), getDefaultArtifactOption(scanOptions))
 	if err != nil {
-		return nil, &os.PathError{Op: "open", Path: name, Err: err}
-	}
-	return &file{File: f, ofs: ofs, fi: fi, name: name}, nil
-}
-
-func (ofs filesystem) path(layerIndex int, name string) string {
-	if !fs.ValidPath(name) {
-		panic(fmt.Errorf("unexpected invalid path name %q", name))
-	}
-	return path.Join(ofs.layers[layerIndex], name)
-}
-
-func (ofs filesystem) stat(name string) (int, fs.FileInfo, error) {
-	var errf error
-	for layerIndex := range ofs.layers {
-		fi, err := os.Stat(ofs.path(layerIndex, name))
-		if errors.Is(err, syscall.ENOENT) || errors.Is(err, syscall.ENOTDIR) {
-			// When path does not exist, overlayfs does not verify that a
-			// whiteout file has been created as one of the parent dir in the
-			// current layer. Meaning you can open file from lower dirs even
-			// if a whiteout or opaque directory has been created on an upper
-			// layer.
-			continue
-		}
-		if err != nil {
-			errf = err
-			break
-		}
-		if isWhiteout(fi) {
-			break
-		}
-		return layerIndex, fi, nil
-	}
-	if errf == nil {
-		errf = syscall.ENOENT
-	}
-	return 0, nil, &os.PathError{Op: "stat", Path: name, Err: errf}
-}
-
-// ReadDir implements fs.ReadDirFS.
-func (ofs filesystem) ReadDir(name string) ([]fs.DirEntry, error) {
-	return ofs.readDirN(name, -1)
-}
-
-func (ofs filesystem) readDirN(name string, n int) ([]fs.DirEntry, error) {
-	name = path.Join("/", name)[1:]
-	if name == "" {
-		name = "."
+		return nil, err
 	}
 
-	var entriesMap map[string]*fs.DirEntry
-	var err error
-	var ok bool
-	for layerIndex := range ofs.layers {
-		if ok, err = ofs.readDirLayer(layerIndex, name, n, &entriesMap); ok {
-			break
-		}
-	}
-	if err == nil && entriesMap == nil {
-		err = syscall.ENOENT
-	}
+	log.Debugf("Generating SBOM for image %s using overlayfs %+v", imgMeta.ID, layers)
+
+	trivyReport, err := c.scan(ctx, containerArtifact, applier.NewApplier(cache), imgMeta, cache, false)
 	if err != nil {
-		return []fs.DirEntry{}, &os.PathError{Op: "readdirent", Path: name, Err: err}
-	}
-
-	entries := make([]fs.DirEntry, 0, len(entriesMap))
-	for _, entry := range entriesMap {
-		if entry != whiteout {
-			entries = append(entries, *entry)
+		if imgMeta != nil {
+			return nil, fmt.Errorf("unable to marshal report to sbom format for image %s, err: %w", imgMeta.ID, err)
 		}
-	}
-	sort.Slice(entries, func(i, j int) bool {
-		return entries[i].Name() < entries[j].Name()
-	})
-	if n > 0 && len(entries) > n {
-		entries = entries[:n]
-	}
-	return entries, nil
-}
-
-func (ofs filesystem) readDirLayer(layerIndex int, name string, n int, entriesMap *map[string]*fs.DirEntry) (bool, error) {
-	fullname := ofs.path(layerIndex, name)
-
-	di, err := os.Stat(fullname)
-	if errors.Is(err, syscall.ENOENT) || errors.Is(err, syscall.ENOTDIR) {
-		return false, nil
-	}
-	if err != nil {
-		return true, err
-	}
-	if isWhiteout(di) {
-		return true, syscall.ENOENT
-	}
-	if !di.IsDir() {
-		return true, syscall.ENOTDIR
+		return nil, fmt.Errorf("unable to marshal report to sbom format, err: %w", err)
 	}
 
-	d, err := os.Open(fullname)
-	if err != nil {
-		return true, err
+	log.Debugf("Found OS: %+v", trivyReport.Metadata.OS)
+	pkgCount := 0
+	for _, results := range trivyReport.Results {
+		pkgCount += len(results.Packages)
 	}
+	log.Debugf("Found %d packages", pkgCount)
 
-	entries, err := d.ReadDir(n)
-	if err != nil {
-		return true, err
-	}
-	if *entriesMap == nil {
-		*entriesMap = make(map[string]*fs.DirEntry)
-	}
-	for entryIndex, entry := range entries {
-		entryName := entry.Name()
-		if _, exists := (*entriesMap)[entryName]; !exists {
-			entryPtr := &entries[entryIndex]
-			if entry.Type().IsRegular() {
-				(*entriesMap)[entryName] = entryPtr
-			} else {
-				ei, err := entry.Info()
-				if err != nil {
-					return true, err
-				}
-				if isWhiteout(ei) {
-					(*entriesMap)[entryName] = whiteout
-				} else {
-					(*entriesMap)[entryName] = entryPtr
-				}
-			}
-		}
-	}
-
-	return isOpaqueDir(d), nil
-}
-
-// ReadDir implements fs.ReadDirFile.
-func (f *file) ReadDir(n int) ([]fs.DirEntry, error) {
-	if !f.fi.IsDir() {
-		return nil, &os.PathError{Op: "readdirent", Path: f.name, Err: syscall.ENOTDIR}
-	}
-	return f.ofs.readDirN(f.name, n)
-}
-
-// Read implements fs.File.
-func (f *file) Read(b []byte) (int, error) {
-	return f.File.Read(b)
-}
-
-// Stat implements fs.File.
-func (f *file) Stat() (fs.FileInfo, error) {
-	return f.fi, nil
-}
-
-// Close implements fs.File.
-func (f *file) Close() error {
-	return f.File.Close()
-}
-
-var _ fs.ReadDirFile = &file{}
-
-func isWhiteout(fm fs.FileInfo) bool {
-	return fm.Mode()&fs.ModeCharDevice != 0 && uint64(fm.Sys().(*syscall.Stat_t).Rdev) == whiteoutCharDev
-}
-
-func isOpaqueDir(d *os.File) bool {
-	var data [1]byte
-	var sz int
-	var err error
-	for {
-		sz, err = unix.Fgetxattr(int(d.Fd()), "trusted.overlay.opaque", data[:])
-		if err != unix.EINTR {
-			break
-		}
-	}
-	return sz == 1 && data[0] == 'y'
+	return &Report{
+		Report:    trivyReport,
+		id:        imgMeta.ID,
+		marshaler: c.marshaler,
+	}, nil
 }
